ft: registration page
 - setup nextjs
 - setup typescript with nextjs
 - setup tailwindcss
 - design registration page
