+++ client/package.json
@@ -8,8 +8,16 @@
     "start": "next start"
   },
   "dependencies": {
-    "next": "10.1.3",
+    "next": "^10.1.3",
     "react": "17.0.2",
     "react-dom": "17.0.2"
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.37",
+    "@types/react": "^17.0.3",
+    "autoprefixer": "^10.2.5",
+    "postcss": "^8.2.10",
+    "tailwindcss": "^2.1.1",
+    "typescript": "^4.2.4"
   }
 }
