+++ client/src/pages/_app.tsx
@@ -1,0 +1,9 @@
+import { AppProps } from 'next/app'
+
+import '../styles/globals.css'
+
+function App({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />
+}
+
+export default App
