+++ client/src/styles/globals.css
@@ -1,0 +1,4 @@
+/* ./styles/globals.css */
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
